+++ src/connectionMatrix/model.cpp
@@ -34,6 +34,7 @@ class HeaderItem : public QStandardItem
 {
 public:
 	using StreamMap = std::unordered_map<la::avdecc::entity::model::StreamIndex, std::int32_t>;
+	using InterfaceMap = std::unordered_map<la::avdecc::entity::model::AvbInterfaceIndex, std::vector<std::int32_t>>;
 	using RelativeParentIndex = std::optional<std::int32_t>;
 
 	HeaderItem(Model::NodeType const nodeType, la::avdecc::UniqueIdentifier const& entityID)
@@ -52,14 +53,20 @@ public:
 		return _entityID;
 	}
 
-	void setStreamIndex(la::avdecc::entity::model::StreamIndex const streamIndex)
+	void setStreamNodeInfo(la::avdecc::entity::model::StreamIndex const streamIndex, la::avdecc::entity::model::AvbInterfaceIndex const avbInterfaceIndex)
 	{
 		_streamIndex = streamIndex;
+		_avbInterfaceIndex = avbInterfaceIndex;
 	}
 
 	la::avdecc::entity::model::StreamIndex streamIndex() const
 	{
 		return _streamIndex;
+	}
+
+	la::avdecc::entity::model::AvbInterfaceIndex avbInterfaceIndex() const
+	{
+		return _avbInterfaceIndex;
 	}
 
 	void setRedundantIndex(la::avdecc::controller::model::VirtualIndex const redundantIndex)
@@ -110,6 +117,16 @@ public:
 	StreamMap const& streamMap() const
 	{
 		return _streamMap;
+	}
+
+	void setInterfaceMap(InterfaceMap const& interfaceMap)
+	{
+		_interfaceMap = interfaceMap;
+	}
+
+	InterfaceMap const& interfaceMap() const
+	{
+		return _interfaceMap;
 	}
 
 	virtual QVariant data(int role) const override
@@ -206,12 +223,14 @@ public:
 private:
 	Model::NodeType const _nodeType;
 	la::avdecc::UniqueIdentifier const _entityID;
-	la::avdecc::entity::model::StreamIndex _streamIndex{ static_cast<la::avdecc::entity::model::StreamIndex>(-1) };
-	la::avdecc::controller::model::VirtualIndex _redundantIndex{ static_cast<la::avdecc::controller::model::VirtualIndex>(-1) };
+	la::avdecc::entity::model::StreamIndex _streamIndex{ la::avdecc::entity::model::getInvalidDescriptorIndex() };
+	la::avdecc::entity::model::AvbInterfaceIndex _avbInterfaceIndex{ la::avdecc::entity::model::getInvalidDescriptorIndex() };
+	la::avdecc::controller::model::VirtualIndex _redundantIndex{ la::avdecc::entity::model::getInvalidDescriptorIndex() };
 	std::int32_t _redundantStreamOrder{ -1 };
 	RelativeParentIndex _relativeParentIndex{ std::nullopt };
 	std::int32_t _childrenCount{ 0 };
 	StreamMap _streamMap{};
+	InterfaceMap _interfaceMap{};
 };
 
 Model::ConnectionCapabilities computeConnectionCapabilities(HeaderItem const* talkerItem, HeaderItem const* listenerItem)
@@ -256,6 +275,16 @@ Model::ConnectionCapabilities computeConnectionCapabilities(HeaderItem const* ta
 			{
 				try
 				{
+					// First check the link status
+					auto const talkerLinkStatus = talkerEntity->getAvbInterfaceLinkStatus(talkerAvbInterfaceIndex);
+					auto const listenerLinkStatus = listenerEntity->getAvbInterfaceLinkStatus(listenerAvbInterfaceIndex);
+
+					// If either is Down, no need to check the domain, it makes no sense (so return compatible)
+					if (talkerLinkStatus == la::avdecc::controller::ControlledEntity::InterfaceLinkStatus::Down || listenerLinkStatus == la::avdecc::controller::ControlledEntity::InterfaceLinkStatus::Down)
+					{
+						return true;
+					}
+
 					// Get the AvbInterface associated to the streams
 					auto const& talkerAvbInterfaceNode = talkerEntity->getAvbInterfaceNode(talkerEntityNode.dynamicModel->currentConfiguration, talkerAvbInterfaceIndex);
 					auto const& listenerAvbInterfaceNode = listenerEntity->getAvbInterfaceNode(listenerEntityNode.dynamicModel->currentConfiguration, listenerAvbInterfaceIndex);
@@ -276,12 +305,22 @@ Model::ConnectionCapabilities computeConnectionCapabilities(HeaderItem const* ta
 				Connected,
 			};
 
-			auto const computeCapabilities = [](ConnectState const connectState, bool const areAllConnected, bool const isFormatCompatible, bool const isDomainCompatible)
+			auto const computeCapabilities = [](bool const interfaceDown, ConnectState const connectState, bool const areAllConnected, bool const isFormatCompatible, bool const isDomainCompatible)
 			{
-				auto caps{ Model::ConnectionCapabilities::Connectable };
+				auto caps{ Model::ConnectionCapabilities::Connectable }; // If we get to this function, we are at least connectable
 
-				if (!isDomainCompatible)
-					caps |= Model::ConnectionCapabilities::WrongDomain;
+				if (interfaceDown)
+				{
+					caps |= Model::ConnectionCapabilities::InterfaceDown;
+				}
+				else
+				{
+					// We can only check if domain is compatible when interface is up (makes no sense otherwise)
+					if (!isDomainCompatible)
+					{
+						caps |= Model::ConnectionCapabilities::WrongDomain;
+					}
+				}
 
 				if (!isFormatCompatible)
 					caps |= Model::ConnectionCapabilities::WrongFormat;
@@ -299,8 +338,14 @@ Model::ConnectionCapabilities computeConnectionCapabilities(HeaderItem const* ta
 				return caps;
 			};
 
-			// Special case for both redundant nodes
-			if (talkerNodeType == Model::NodeType::RedundantOutput && listenerNodeType == Model::NodeType::RedundantInput)
+			// At least one entity node: we want to know if at least one connection is established
+			if (talkerNodeType == Model::NodeType::Entity || listenerNodeType == Model::NodeType::Entity)
+			{
+				// TODO
+				return computeCapabilities(false, ConnectState::NotConnected, false, false, false);
+			}
+			// Both redundant nodes: we want to differentiate full redundant connection (both pairs connected) from partial one (only one of the pair connected)
+			else if (talkerNodeType == Model::NodeType::RedundantOutput && listenerNodeType == Model::NodeType::RedundantInput)
 			{
 				// Check if all redundant streams are connected
 				auto const& talkerRedundantNode = talkerEntity->getRedundantStreamOutputNode(talkerEntityNode.dynamicModel->currentConfiguration, talkerRedundantIndex);
@@ -311,6 +356,7 @@ Model::ConnectionCapabilities computeConnectionCapabilities(HeaderItem const* ta
 				auto const* const talkerStreamNode = static_cast<la::avdecc::controller::model::StreamOutputNode const*>(talkerIt->second);
 				auto listenerIt = listenerRedundantNode.redundantStreams.begin();
 				auto const* const listenerStreamNode = static_cast<la::avdecc::controller::model::StreamInputNode const*>(listenerIt->second);
+				auto atLeastOneInterfaceDown{ false };
 				auto atLeastOneConnected{ false };
 				auto allConnected{ true };
 				auto allCompatibleFormat{ true };
@@ -320,6 +366,7 @@ Model::ConnectionCapabilities computeConnectionCapabilities(HeaderItem const* ta
 					auto const* const redundantTalkerStreamNode = static_cast<la::avdecc::controller::model::StreamOutputNode const*>(talkerIt->second);
 					auto const* const redundantListenerStreamNode = static_cast<la::avdecc::controller::model::StreamInputNode const*>(listenerIt->second);
 					auto const connected = avdecc::helper::isStreamConnected(talkerEntityID, redundantTalkerStreamNode, redundantListenerStreamNode);
+					atLeastOneInterfaceDown |= (talkerEntity->getAvbInterfaceLinkStatus(redundantTalkerStreamNode->staticModel->avbInterfaceIndex) == la::avdecc::controller::ControlledEntity::InterfaceLinkStatus::Down) || (listenerEntity->getAvbInterfaceLinkStatus(redundantListenerStreamNode->staticModel->avbInterfaceIndex) == la::avdecc::controller::ControlledEntity::InterfaceLinkStatus::Down);
 					atLeastOneConnected |= connected;
 					allConnected &= connected;
 					allCompatibleFormat &= computeFormatCompatible(*redundantTalkerStreamNode, *redundantListenerStreamNode);
@@ -328,9 +375,15 @@ Model::ConnectionCapabilities computeConnectionCapabilities(HeaderItem const* ta
 					++listenerIt;
 				}
 
-				return computeCapabilities(atLeastOneConnected ? ConnectState::Connected : ConnectState::NotConnected, allConnected, allCompatibleFormat, allDomainCompatible);
+				return computeCapabilities(atLeastOneInterfaceDown, atLeastOneConnected ? ConnectState::Connected : ConnectState::NotConnected, allConnected, allCompatibleFormat, allDomainCompatible);
 			}
-			else if ((talkerNodeType == Model::NodeType::OutputStream && listenerNodeType == Model::NodeType::InputStream) || (talkerNodeType == Model::NodeType::RedundantOutputStream && listenerNodeType == Model::NodeType::RedundantInputStream) || (talkerNodeType == Model::NodeType::RedundantOutput && listenerNodeType == Model::NodeType::RedundantInputStream) || (talkerNodeType == Model::NodeType::RedundantOutputStream && listenerNodeType == Model::NodeType::RedundantInput))
+			// One non-redundant stream and one redundant node: We want to check if one connection is active or possible (only one should be, a non-redundant device can only be connected with either of the redundant domain pair)
+			else if ((talkerNodeType == Model::NodeType::OutputStream && listenerNodeType == Model::NodeType::RedundantInput) || (talkerNodeType == Model::NodeType::RedundantOutput && listenerNodeType == Model::NodeType::InputStream))
+			{
+				return computeCapabilities(false, ConnectState::FastConnecting, false, false, false);
+			}
+			// All other cases: There is only one connection possibility
+			else
 			{
 				la::avdecc::controller::model::StreamOutputNode const* talkerNode{ nullptr };
 				la::avdecc::controller::model::StreamInputNode const* listenerNode{ nullptr };
@@ -377,6 +430,7 @@ Model::ConnectionCapabilities computeConnectionCapabilities(HeaderItem const* ta
 				}
 
 				// Get connected state
+				auto const interfaceDown = (talkerEntity->getAvbInterfaceLinkStatus(talkerNode->staticModel->avbInterfaceIndex) == la::avdecc::controller::ControlledEntity::InterfaceLinkStatus::Down) || (listenerEntity->getAvbInterfaceLinkStatus(listenerNode->staticModel->avbInterfaceIndex) == la::avdecc::controller::ControlledEntity::InterfaceLinkStatus::Down);
 				auto const areConnected = avdecc::helper::isStreamConnected(talkerEntityID, talkerNode, listenerNode);
 				auto const fastConnecting = avdecc::helper::isStreamFastConnecting(talkerEntityID, talkerNode, listenerNode);
 				auto const connectState = areConnected ? ConnectState::Connected : (fastConnecting ? ConnectState::FastConnecting : ConnectState::NotConnected);
@@ -387,7 +441,7 @@ Model::ConnectionCapabilities computeConnectionCapabilities(HeaderItem const* ta
 				// Get domain compatibility
 				auto const isDomainCompatible = computeDomainCompatible(talkerNode->staticModel->avbInterfaceIndex, listenerNode->staticModel->avbInterfaceIndex);
 
-				return computeCapabilities(connectState, areConnected, isFormatCompatible, isDomainCompatible);
+				return computeCapabilities(interfaceDown, connectState, areConnected, isFormatCompatible, isDomainCompatible);
 			}
 		}
 	}
@@ -443,6 +497,7 @@ public:
 		connect(&controllerManager, &avdecc::ControllerManager::gptpChanged, this, &ModelPrivate::gptpChanged);
 		connect(&controllerManager, &avdecc::ControllerManager::entityNameChanged, this, &ModelPrivate::entityNameChanged);
 		connect(&controllerManager, &avdecc::ControllerManager::streamNameChanged, this, &ModelPrivate::streamNameChanged);
+		connect(&controllerManager, &avdecc::ControllerManager::avbInterfaceLinkStatusChanged, this, &ModelPrivate::avbInterfaceLinkStatusChanged);
 	}
 
 	// Slots for avdecc::ControllerManager signals
@@ -475,8 +530,9 @@ public:
 
 				if (la::avdecc::hasFlag(controlledEntity->getEntity().getTalkerCapabilities(), la::avdecc::entity::TalkerCapabilities::Implemented) && !configurationNode.streamOutputs.empty())
 				{
-					std::int32_t streamMapIndex{ 0 };
+					std::int32_t offsetFromEntityNode{ 0 };
 					HeaderItem::StreamMap streamMap{};
+					HeaderItem::InterfaceMap interfaceMap{};
 
 					std::int32_t const entityItemIndex{ q_ptr->rowCount() };
 					std::int32_t entityItemChildrenCount{ 0 };
@@ -490,7 +546,7 @@ public:
 						std::int32_t const redundantItemIndex{ q_ptr->rowCount() };
 						std::int32_t redundantItemChildrenCount{ 0 };
 
-						auto const& redundantIndex{ output.first };
+						auto const redundantIndex{ output.first };
 						auto const& redundantNode{ output.second };
 
 						auto* redundantItem = new HeaderItem(Model::NodeType::RedundantOutput, entityID);
@@ -499,19 +555,23 @@ public:
 						q_ptr->setVerticalHeaderItem(redundantItemIndex, redundantItem);
 
 						++entityItemChildrenCount;
-						++streamMapIndex;
+						++offsetFromEntityNode;
 
 						std::int32_t redundantStreamOrder{ 0 };
 						for (auto const& streamKV : redundantNode.redundantStreams)
 						{
 							std::int32_t const redundantStreamItemIndex{ q_ptr->rowCount() };
 
-							auto const& streamIndex{ streamKV.first };
-							streamMap.insert(std::make_pair(streamIndex, ++streamMapIndex));
+							auto const streamIndex{ streamKV.first };
+							auto const interfaceIndex{ streamKV.second->staticModel->avbInterfaceIndex };
+							auto const currentOffset{ ++offsetFromEntityNode };
+							streamMap.insert(std::make_pair(streamIndex, currentOffset));
+							auto& mapIndexes = interfaceMap[interfaceIndex];
+							mapIndexes.push_back(currentOffset);
 
 							auto* redundantStreamItem = new HeaderItem(Model::NodeType::RedundantOutputStream, entityID);
 							redundantStreamItem->setRelativeParentIndex(redundantItemIndex - redundantStreamItemIndex);
-							redundantStreamItem->setStreamIndex(streamIndex);
+							redundantStreamItem->setStreamNodeInfo(streamIndex, interfaceIndex);
 							redundantStreamItem->setRedundantIndex(redundantIndex);
 							redundantStreamItem->setRedundantStreamOrder(redundantStreamOrder);
 							q_ptr->setVerticalHeaderItem(redundantStreamItemIndex, redundantStreamItem);
@@ -528,17 +588,21 @@ public:
 					// Single streams
 					for (auto const& output : configurationNode.streamOutputs)
 					{
-						auto const& streamIndex{ output.first };
+						auto const streamIndex{ output.first };
 						auto const& streamNode{ output.second };
 
 						if (!streamNode.isRedundant)
 						{
 							std::int32_t const streamItemIndex{ q_ptr->rowCount() };
-							streamMap.insert(std::make_pair(streamIndex, ++streamMapIndex));
+							auto const interfaceIndex{ streamNode.staticModel->avbInterfaceIndex };
+							auto const currentOffset{ ++offsetFromEntityNode };
+							streamMap.insert(std::make_pair(streamIndex, currentOffset));
+							auto& mapIndexes = interfaceMap[interfaceIndex];
+							mapIndexes.push_back(currentOffset);
 
 							auto* streamItem = new HeaderItem{ Model::NodeType::OutputStream, entityID };
 							streamItem->setRelativeParentIndex(entityItemIndex - streamItemIndex);
-							streamItem->setStreamIndex(streamIndex);
+							streamItem->setStreamNodeInfo(streamIndex, streamNode.staticModel->avbInterfaceIndex);
 							q_ptr->setVerticalHeaderItem(streamItemIndex, streamItem);
 
 							++entityItemChildrenCount;
@@ -547,6 +611,7 @@ public:
 
 					entityItem->setChildrenCount(entityItemChildrenCount);
 					entityItem->setStreamMap(streamMap);
+					entityItem->setInterfaceMap(interfaceMap);
 
 					// Create new connection items
 					for (auto column = 0; column < q_ptr->columnCount(); ++column)
@@ -565,8 +630,9 @@ public:
 
 				if (la::avdecc::hasFlag(controlledEntity->getEntity().getListenerCapabilities(), la::avdecc::entity::ListenerCapabilities::Implemented) && !configurationNode.streamInputs.empty())
 				{
-					std::int32_t streamMapIndex{ 0 };
+					std::int32_t offsetFromEntityNode{ 0 };
 					HeaderItem::StreamMap streamMap{};
+					HeaderItem::InterfaceMap interfaceMap{};
 
 					std::int32_t const entityItemIndex{ q_ptr->columnCount() };
 					std::int32_t entityItemChildrenCount{ 0 };
@@ -580,7 +646,7 @@ public:
 						std::int32_t const redundantItemIndex{ q_ptr->columnCount() };
 						std::int32_t redundantItemChildrenCount{ 0 };
 
-						auto const& redundantIndex{ input.first };
+						auto const redundantIndex{ input.first };
 						auto const& redundantNode{ input.second };
 
 						auto* redundantItem = new HeaderItem(Model::NodeType::RedundantInput, entityID);
@@ -589,19 +655,23 @@ public:
 						q_ptr->setHorizontalHeaderItem(redundantItemIndex, redundantItem);
 
 						++entityItemChildrenCount;
-						++streamMapIndex;
+						++offsetFromEntityNode;
 
 						std::int32_t redundantStreamOrder{ 0 };
 						for (auto const& streamKV : redundantNode.redundantStreams)
 						{
 							std::int32_t const redundantStreamItemIndex{ q_ptr->columnCount() };
 
-							auto const& streamIndex{ streamKV.first };
-							streamMap.insert(std::make_pair(streamIndex, ++streamMapIndex));
+							auto const streamIndex{ streamKV.first };
+							auto const interfaceIndex{ streamKV.second->staticModel->avbInterfaceIndex };
+							auto const currentOffset{ ++offsetFromEntityNode };
+							streamMap.insert(std::make_pair(streamIndex, currentOffset));
+							auto& mapIndexes = interfaceMap[interfaceIndex];
+							mapIndexes.push_back(currentOffset);
 
 							auto* redundantStreamItem = new HeaderItem(Model::NodeType::RedundantInputStream, entityID);
 							redundantStreamItem->setRelativeParentIndex(redundantItemIndex - redundantStreamItemIndex);
-							redundantStreamItem->setStreamIndex(streamIndex);
+							redundantStreamItem->setStreamNodeInfo(streamIndex, streamKV.second->staticModel->avbInterfaceIndex);
 							redundantStreamItem->setRedundantIndex(redundantIndex);
 							redundantStreamItem->setRedundantStreamOrder(redundantStreamOrder);
 							q_ptr->setHorizontalHeaderItem(redundantStreamItemIndex, redundantStreamItem);
@@ -618,17 +688,21 @@ public:
 					// Single streams
 					for (auto const& input : configurationNode.streamInputs)
 					{
-						auto const& streamIndex{ input.first };
+						auto const streamIndex{ input.first };
 						auto const& streamNode{ input.second };
 
 						if (!streamNode.isRedundant)
 						{
 							std::int32_t const streamItemIndex{ q_ptr->columnCount() };
-							streamMap.insert(std::make_pair(streamIndex, ++streamMapIndex));
+							auto const interfaceIndex{ streamNode.staticModel->avbInterfaceIndex };
+							auto const currentOffset{ ++offsetFromEntityNode };
+							streamMap.insert(std::make_pair(streamIndex, currentOffset));
+							auto& mapIndexes = interfaceMap[interfaceIndex];
+							mapIndexes.push_back(currentOffset);
 
 							auto* streamItem = new HeaderItem{ Model::NodeType::InputStream, entityID };
 							streamItem->setRelativeParentIndex(entityItemIndex - streamItemIndex);
-							streamItem->setStreamIndex(streamIndex);
+							streamItem->setStreamNodeInfo(streamIndex, streamNode.staticModel->avbInterfaceIndex);
 							q_ptr->setHorizontalHeaderItem(streamItemIndex, streamItem);
 
 							++entityItemChildrenCount;
@@ -637,6 +711,7 @@ public:
 
 					entityItem->setChildrenCount(entityItemChildrenCount);
 					entityItem->setStreamMap(streamMap);
+					entityItem->setInterfaceMap(interfaceMap);
 
 					// Create new connection cells
 					for (auto row = 0; row < q_ptr->rowCount(); ++row)
@@ -747,8 +822,32 @@ public:
 		}
 	}
 
+	Q_SLOT void avbInterfaceLinkStatusChanged(la::avdecc::UniqueIdentifier const entityID, la::avdecc::entity::model::AvbInterfaceIndex const avbInterfaceIndex, la::avdecc::controller::ControlledEntity::InterfaceLinkStatus const linkStatus)
+	{
+		LOG_HIVE_DEBUG(QString("connectionMatrix::Model::avbInterfaceLinkStatusChanged: EntityID=%1 Index=%2").arg(avdecc::helper::uniqueIdentifierToString(entityID)).arg(avbInterfaceIndex));
+
+		// Get talker indexes using this AVB Interface
+		{
+			auto const indexes = talkerInterfaceIndexes(entityID, avbInterfaceIndex);
+			for (auto const& index : indexes)
+			{
+				dataChanged(index, true, false);
+			}
+		}
+
+		// Get listener indexes using this AVB Interface
+		{
+			auto const indexes = listenerInterfaceIndexes(entityID, avbInterfaceIndex);
+			for (auto const& index : indexes)
+			{
+				dataChanged(index, true, false);
+			}
+		}
+	}
+
 	QModelIndex talkerIndex(la::avdecc::UniqueIdentifier const entityID) const
 	{
+#pragma message("TODO: Optimization: Build and update (on entity online/offline events) an unordered_map that stores the index of the Entity")
 		for (auto row = 0; row < q_ptr->rowCount(); ++row)
 		{
 			auto* item = static_cast<HeaderItem*>(q_ptr->verticalHeaderItem(row));
@@ -785,15 +884,42 @@ public:
 		return {};
 	}
 
+	QModelIndexList talkerInterfaceIndexes(la::avdecc::UniqueIdentifier const entityID, la::avdecc::entity::model::AvbInterfaceIndex const avbInterfaceIndex) const
+	{
+		try
+		{
+			for (auto row = 0; row < q_ptr->rowCount(); ++row)
+			{
+				auto* item = static_cast<HeaderItem*>(q_ptr->verticalHeaderItem(row));
+				if (item->nodeType() == Model::NodeType::Entity && item->entityID() == entityID)
+				{
+					auto const& interfaceMap = item->interfaceMap();
+					auto const& offsets = interfaceMap.at(avbInterfaceIndex);
+					auto indexes = QModelIndexList{};
+					for (auto const offset : offsets)
+					{
+						indexes.push_back(q_ptr->createIndex(row + offset, -1));
+					}
+					return indexes;
+				}
+			}
+		}
+		catch (std::out_of_range const&)
+		{
+			// Something went wrong and .at() throw
+			LOG_HIVE_ERROR(QString("connectionMatrix::Model::talkerInterfaceIndex: Invalid AvbInterfaceIndex: TalkerID=%1 Index=%2 RowCount=%3 ").arg(avdecc::helper::uniqueIdentifierToString(entityID)).arg(avbInterfaceIndex).arg(q_ptr->rowCount()));
+		}
+
+		return {};
+	}
+
 	QModelIndex listenerIndex(la::avdecc::UniqueIdentifier const entityID) const
 	{
 		for (auto column = 0; column < q_ptr->columnCount(); ++column)
 		{
-			// Refresh whole rows for specified talker
 			auto* item = static_cast<HeaderItem*>(q_ptr->horizontalHeaderItem(column));
 			if (item->nodeType() == Model::NodeType::Entity && item->entityID() == entityID)
 			{
-				// Refresh whole columns for specified listener
 				return q_ptr->createIndex(-1, column);
 			}
 		}
@@ -807,11 +933,9 @@ public:
 		{
 			for (auto column = 0; column < q_ptr->columnCount(); ++column)
 			{
-				// Refresh talker header
 				auto* item = static_cast<HeaderItem*>(q_ptr->horizontalHeaderItem(column));
 				if (item->nodeType() == Model::NodeType::Entity && item->entityID() == entityID)
 				{
-					// Refresh listener header
 					auto const& streamMap = item->streamMap();
 					auto const offset = streamMap.at(streamIndex);
 					return q_ptr->createIndex(-1, column + offset);
@@ -822,6 +946,35 @@ public:
 		{
 			// Something went wrong and .at() throw
 			LOG_HIVE_ERROR(QString("connectionMatrix::Model::listenerStreamIndex: Invalid StreamIndex: ListenerID=%1 Index=%2 ColumnCount=%3 ").arg(avdecc::helper::uniqueIdentifierToString(entityID)).arg(streamIndex).arg(q_ptr->columnCount()));
+		}
+
+		return {};
+	}
+
+	QModelIndexList listenerInterfaceIndexes(la::avdecc::UniqueIdentifier const entityID, la::avdecc::entity::model::AvbInterfaceIndex const avbInterfaceIndex) const
+	{
+		try
+		{
+			for (auto column = 0; column < q_ptr->columnCount(); ++column)
+			{
+				auto* item = static_cast<HeaderItem*>(q_ptr->horizontalHeaderItem(column));
+				if (item->nodeType() == Model::NodeType::Entity && item->entityID() == entityID)
+				{
+					auto const& interfaceMap = item->interfaceMap();
+					auto const& offsets = interfaceMap.at(avbInterfaceIndex);
+					auto indexes = QModelIndexList{};
+					for (auto const offset : offsets)
+					{
+						indexes.push_back(q_ptr->createIndex(-1, column + offset));
+					}
+					return indexes;
+				}
+			}
+		}
+		catch (std::out_of_range const&)
+		{
+			// Something went wrong and .at() throw
+			LOG_HIVE_ERROR(QString("connectionMatrix::Model::listenerInterfaceIndex: Invalid AvbInterfaceIndex: ListenerID=%1 Index=%2 ColumnCount=%3 ").arg(avdecc::helper::uniqueIdentifierToString(entityID)).arg(avbInterfaceIndex).arg(q_ptr->columnCount()));
 		}
 
 		return {};
@@ -887,6 +1040,35 @@ public:
 		}
 	}
 
+#pragma message("TODO: Rework how updateIntersectionCapabilities() is computed, see the following note")
+	/*
+	Fully rework how dataChanged is used to update the intersection data:
+	 - Have an EnumBitfield with the following bits, that is passed to the dataChanged method (so we don't recompute everything when only the format changes for example):
+	   - UpdateConnectable: Update the connectable state of the intersection (should only be called once during first computation, the connectable state never changes)
+		 - UpdateConnected: Update the connected status, or the summary if this is a parent node
+		 - UpdateFormat: Update the matching format status, or the summary if this is a parent node
+		 - UpdateGptp: Update the matching gPTP status, or the summary if this is a parent node
+		 - UpdateLinkStatus: Update the link status, or the summary if this is a parent node
+
+	 - Rename ConnectionCapabilitiesRole to IntersectionCapabilitiesRole (better reflect that it's the intersection, not just the connection: might not be connectable)
+	 - Add new roles:
+	   - PrimaryChildConnectionCapabilitiesRole: Returns the ConnectionCapabilities of the primary child (only valid for the intersection of 2 RedundantNodes), useful to display detailled error
+		 - SecondaryChildConnectionCapabilitiesRole: Returns the ConnectionCapabilities of the secondary child (only valid for the intersection of 2 RedundantNodes), useful to display detailled error
+	 - Remove ConnectionCapabilities::PartiallyConnected (no longer required, PrimaryChildConnectionCapabilitiesRole and SecondaryChildConnectionCapabilitiesRole should be used instead)
+	 - Change ConnectionCapabilities::InterfaceDown so it only return the status for a valid stream, not the redundant summary (shoud use PrimaryChildConnectionCapabilitiesRole and SecondaryChildConnectionCapabilitiesRole instead)
+
+	This should achieve better performance because we don't have to undergo the complete updateIntersectionCapabilities method everytime a single thing change.
+	Then the itemDelegate.cpp:paint() method should be much more simplier:
+	 - Check for the symbol to draw:
+	   - IntersectionConnectableRole is false -> Empty
+	   - isEntityCrossSection (to be computed based on NodeTypeRole) -> Square
+	   - At least one of the 2 is a redundantStream (based on NodeTypeRole) -> Lozenge
+	   - Else -> Circle
+	 - Then get the color to draw:
+	   - IntersectionConnectedRole -> Dark or Light
+		 - IntersectionFormatRole, IntersectionGptpRole and IntersectionLinkStatusRole gets the error status -> The view can actually choose what error to display before the other
+	Always recompute childs first, so that parents can assume the data of each child is up-to-date to build the summary (instead of having to call avdecc methods again)
+	*/
 	void dataChanged(QModelIndex const& index, bool const andParents, bool const andChildren)
 	{
 		if (index.row() == -1 && index.column() == -1)