+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 - Icon when an entity is acquired by Hive (changed color from orange to green)
+- Changed the colors in the Connection Matrix (see Legend)
 
 ### Fixed
 - Upload firmware progression always set to 100% upon successfull completion